feat[judge-handler]: add submission-related methods

+++ src/main/java/com/example/codeE/model/exercise/common/SubmissionTestCase.java
@@ -32,16 +32,16 @@ public class SubmissionTestCase {
     private String status;
 
     @Field("time")
-    private Float time;
+    private Double time;
 
     @Field("memory")
-    private Float memory;
+    private Double memory;
 
     @Field("points")
-    private Float points;
+    private Double points;
 
     @Field("total")
-    private Float total;
+    private Double total;
 
     @Field("feedback")
     private String feedback;
